editoast: remove models module
