+++ editoast/editoast_derive/src/modelv2/codegen/retrieve_batch_impl.rs
@@ -79,7 +79,7 @@ impl ToTokens for RetrieveBatchImpl {
                     conn: &mut editoast_models::DbConnection,
                     ids: I,
                 ) -> crate::error::Result<C> {
-                    use crate::models::Identifiable;
+                    use crate::modelsv2::Identifiable;
                     use crate::modelsv2::Model;
                     use #table_mod::dsl;
                     use diesel::prelude::*;
