+++ editoast/src/views/mod.rs
@@ -53,7 +53,6 @@ use crate::error::Result;
 use crate::error::{self};
 use crate::generated_data;
 use crate::infra_cache::operation;
-use crate::models;
 use crate::modelsv2;
 use crate::modelsv2::auth::PgAuthDriver;
 use crate::AppState;
@@ -89,7 +88,6 @@ editoast_common::schemas! {
 
     editoast_common::schemas(),
     editoast_schemas::schemas(),
-    models::schemas(),
     modelsv2::schemas(),
     core::schemas(),
     documents::schemas(),
